Move graph.graph to graph.mod
