+++ src/prelude.rs
@@ -1,7 +1,4 @@
 pub use crate::data_structure::{segment_tree::SegmentTree, union_find::UnionFind};
 pub use crate::graph::{
-    adjacency_list::AdjListGraph,
-    edge_impl,
-    graph::{Edge, Graph, WeightedEdge},
-    shortest_path,
+    adjacency_list::AdjListGraph, edge_impl, shortest_path, Edge, Graph, WeightedEdge,
 };
